Introduce ProcessCommandStatus instead of DefaultCommandStatus
